+++ client/package.json
@@ -6,11 +6,13 @@
     "@material-ui/core": "^3.8.3",
     "auth0-js": "^9.9.0",
     "auth0-lock": "^11.13.0",
+    "radium": "^0.25.1",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.3"
+    "react-scripts": "2.1.3",
+    "react-springy-parallax": "^1.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
