+++ client/src/App.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import UserProfile from './Components/UserProfile';
 import ListPage from './Components/ListPage/ListPage';
 import {Route, withRouter} from 'react-router-dom';
-import NavBar from "./NavBar/NavBar"
-import Header from "./Header/Header"
-import Main from "./Main/Main"
+import NavBar from "./NavBar/NavBar";
+import Header from "./Header/Header";
+import Welcome from "./WelcomePage/Welcome";
 import './App.css';
 // import auth0Client from './Auth';
 // import SecuredRoute from ;
@@ -39,9 +39,7 @@ class App extends Component {
     return (
       <div className="App">
         <NavBar />
-        <Header />
-        <UserProfile/>
-        <Main />
+        <Welcome />
         <Route exact path='/callback' component={Callback} />
 
       </div>
