+++ config/config.js
@@ -4,7 +4,7 @@ require("dotenv").config();
 module.exports = {
   development: {
     username: "root",
-    password: "root",
+    password: "MyNewPass",
     database: "giftd_db",
     host: "localhost",
     dialect: "mysql"
